requirements.txt
Flask
openai
python-dotenv
pypdf
python-docx